indian_stock_lstm/requirements.txt
numpy>=1.24
pandas>=2.0
matplotlib>=3.7
seaborn>=0.12
scikit-learn>=1.2
yfinance>=0.2.18
tensorflow-cpu>=2.11
jupyter
streamlit>=1.28
plotly>=5.18
python-dotenv>=1.0